Split review Playwright flows
