+++ E2E/playwright/tests/review.spec.ts
@@ -1,0 +1,97 @@
+import { test, expect } from '@playwright/test';
+import { harnessURL } from './harness-helpers';
+
+test('mock harness shows git review summary for diff flow', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('git diff');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByTestId('review-pane')).toBeVisible();
+  await expect(page.getByTestId('review-summary')).toHaveText('1 file changed, +1 -0');
+  await expect(page.getByTestId('review-file')).toContainText('Sources/App.swift');
+  await expect(page.getByTestId('review-line')).toHaveCount(2);
+  await expect(page.getByTestId('review-line').first()).toContainText('let title = "QuillCode"');
+  await expect(page.getByTestId('tool-card-title')).toHaveText('host.git.diff');
+  await expect(page.getByTestId('tool-card-output')).toContainText('diff --git');
+
+  await page.getByLabel('Review note for Sources/App.swift').fill('Check the exported symbol name');
+  await page.getByRole('button', { name: 'Add note' }).click();
+  await expect(page.getByTestId('review-comment')).toContainText('Check the exported symbol name');
+
+  await page.getByLabel('Line note for Sources/App.swift:1').fill('This is a useful exported constant');
+  await page.getByTestId('review-line-comment-form').first().getByRole('button', { name: 'Add' }).click();
+  await expect(page.getByTestId('review-line-comment')).toContainText('This is a useful exported constant');
+
+  await page.getByLabel('Range note for Sources/App.swift').fill('Keep the title adjacent to the import');
+  await page.getByTestId('review-range-comment-form').getByRole('button', { name: 'Add range note' }).click();
+  const rangeComment = page.getByTestId('review-line-comment').filter({ hasText: 'Lines 1-2' });
+  await expect(rangeComment).toContainText('Keep the title adjacent to the import');
+});
+
+test('mock harness flows apply patch into review diff', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('apply patch to edit file');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByTestId('review-pane')).toBeVisible();
+  await expect(page.getByTestId('review-summary')).toHaveText('1 file changed, +1 -0');
+  await expect(page.getByTestId('review-line').first()).toContainText('let title = "QuillCode"');
+  await expect(page.getByTestId('tool-card-title')).toHaveText([
+    'host.apply_patch',
+    'host.git.diff'
+  ]);
+  await expect(page.getByText('Patch applied. Review the resulting diff below.')).toBeVisible();
+});
+
+test('mock harness stages a changed file from the review pane', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('git diff');
+  await page.getByRole('button', { name: 'Send' }).click();
+  await expect(page.getByTestId('review-pane')).toBeVisible();
+  await expect(page.getByTestId('review-action')).toHaveCount(4);
+
+  await page.getByRole('button', { name: 'Stage', exact: true }).click();
+
+  await expect(page.getByTestId('review-pane')).toHaveCount(0);
+  await expect(page.getByTestId('agent-status')).toHaveText('Idle');
+  await expect(page.getByTestId('tool-card-title')).toHaveText([
+    'host.git.diff',
+    'host.git.stage',
+    'host.git.diff'
+  ]);
+  await expect(page.getByTestId('tool-card-input').nth(1)).toContainText('Sources/App.swift');
+});
+
+test('mock harness stages a single hunk from the review pane', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('git diff');
+  await page.getByRole('button', { name: 'Send' }).click();
+  await expect(page.getByTestId('review-hunk')).toBeVisible();
+  await expect(page.getByTestId('review-hunk-header')).toContainText('@@ -1 +1,2 @@');
+
+  await page.getByRole('button', { name: 'Stage hunk' }).click();
+
+  await expect(page.getByTestId('review-pane')).toHaveCount(0);
+  await expect(page.getByTestId('tool-card-title')).toHaveText([
+    'host.git.diff',
+    'host.git.stage_hunk',
+    'host.git.diff'
+  ]);
+  await expect(page.getByTestId('tool-card-input').nth(1)).toContainText('Sources/App.swift');
+});
+
+test('mock harness commits staged changes in one turn', async ({ page }) => {
+  await page.goto(harnessURL());
+
+  await page.getByLabel('Message').fill('commit these changes with message Add hello file');
+  await page.getByRole('button', { name: 'Send' }).click();
+
+  await expect(page.getByTestId('tool-card-title')).toHaveText('host.git.commit');
+  await expect(page.getByTestId('tool-card-input')).toContainText('Add hello file');
+  await expect(page.getByTestId('tool-card-output')).toContainText('[main abc1234] Add hello file');
+  await expect(page.getByText('Output:\n[main abc1234] Add hello file')).toBeVisible();
+});
